builds/requirements.txt
astroid==2.2.5
boto3==1.9.199
botocore==1.12.199
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-storages==1.7.1
djangorestframework==3.10.1
docutils==0.14
gunicorn==19.9.0
httpie==1.0.2
idna==2.8
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pygments==2.4.2
pylint==2.3.1
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
dj-database-url==0.4.0
